8.1.6.141: Make NBT Communicate System

+++ src/main/java/net/blacklab/lmr/client/entity/EntityLittleMaidForTexSelect.java
@@ -165,13 +165,13 @@ public class EntityLittleMaidForTexSelect extends EntityLiving implements IModel
 	}
 
 	@Override
-	public void setColor(int pColor) {
+	public void setColor(byte pColor) {
 		textureData.setColor(pColor);
 //		color = pColor;
 	}
 
 	@Override
-	public int getColor() {
+	public byte getColor() {
 		return textureData.getColor();
 //		return color;
 	}
